Freezer working with custom recipes
