+++ src/main/java/net/artienia/rubinated_nether/block/entity/FreezerBlockEntity.java
@@ -2,18 +2,25 @@ package net.artienia.rubinated_nether.block.entity;
 
 import net.artienia.rubinated_nether.RubinatedNether;
 import net.artienia.rubinated_nether.mixin.AbstractFurnaceBlockEntityAccessor;
+import net.artienia.rubinated_nether.recipe.FreezingRecipe;
 import net.artienia.rubinated_nether.recipe.ModRecipeTypes;
 import net.artienia.rubinated_nether.screen.FreezerMenu;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.NonNullList;
 import net.minecraft.core.RegistryAccess;
+import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.tags.TagKey;
 import net.minecraft.util.Mth;
+import net.minecraft.world.Containers;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.WorldlyContainer;
 import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.ContainerData;
 import net.minecraft.world.inventory.RecipeHolder;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -26,10 +33,17 @@ import net.minecraft.world.level.ItemLike;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.AbstractFurnaceBlock;
 import net.minecraft.world.level.block.entity.AbstractFurnaceBlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ForgeCapabilities;
+import net.minecraftforge.common.util.LazyOptional;
+import net.minecraftforge.items.IItemHandler;
+import net.minecraftforge.items.ItemStackHandler;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.tags.ITagManager;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.LinkedHashMap;
@@ -38,44 +52,26 @@ import java.util.Optional;
 import java.util.stream.Stream;
 
 
-/**
- * [CODE COPY] - {@link AbstractFurnaceBlockEntity}.<br><br>
- * Certain static methods are copied with minor noted changes, and use accessors for {@link AbstractFurnaceBlockEntity}.
- */
-public class FreezerBlockEntity extends AbstractFurnaceBlockEntity {
+
+public class FreezerBlockEntity extends AbstractFreezerBlockEntity implements MenuProvider {
     private static final Map<Item, Integer> freezingMap = new LinkedHashMap<>();
 
-    private static final int[] SLOTS_FOR_UP = new int[]{0};
-    private static final int[] SLOTS_FOR_DOWN = new int[]{2, 0};
-    private static final int[] SLOTS_FOR_SIDES = new int[]{1};
-    protected ItemStack remainderItem = ItemStack.EMPTY;
-
-
-    public FreezerBlockEntity(BlockEntityType<?> type, BlockPos pos, BlockState state, RecipeType<? extends AbstractCookingRecipe> recipeType) {
-        super(type, pos, state, recipeType);
-    }
-
-    public FreezerBlockEntity(BlockPos pPos, BlockState pBlockState) {
-        super(ModBlockEntities.FREEZER_BE.get(), pPos, pBlockState, ModRecipeTypes.FREEZING.get());
+    public FreezerBlockEntity(BlockPos pos, BlockState state) {
+        super(ModBlockEntityTypes.FREEZER.get(), pos, state, ModRecipeTypes.FREEZING.get());
     }
 
     @Override
-    public Component getDisplayName() {
+    public Component getDefaultName() {
         return Component.translatable("menu." + RubinatedNether.MOD_ID + ".freezer");
     }
 
     @Override
-    protected Component getDefaultName() {
-        return Component.translatable("menu." + RubinatedNether.MOD_ID + ".freezer");
-    }
-
-    @Override
-    protected AbstractContainerMenu createMenu(int id, Inventory playerInventory) {
+    public AbstractContainerMenu createMenu(int id, Inventory playerInventory) {
         return new FreezerMenu(id, playerInventory, this, this.dataAccess);
     }
 
     @Override
-    protected int getBurnDuration(ItemStack fuelStack) {
+    public int getBurnDuration(ItemStack fuelStack) {
         if (fuelStack.isEmpty() || !getFreezingMap().containsKey(fuelStack.getItem())) {
             return 0;
         } else {
@@ -86,7 +82,6 @@ public class FreezerBlockEntity extends AbstractFurnaceBlockEntity {
     public static Map<Item, Integer> getFreezingMap() {
         return freezingMap;
     }
-
 
     public static void addItemFreezingTime(ItemLike itemProvider, int burnTime) {
         Item item = itemProvider.asItem();
@@ -119,165 +114,4 @@ public class FreezerBlockEntity extends AbstractFurnaceBlockEntity {
             tags.getTag(itemTag).stream().forEach((item) -> getFreezingMap().remove(item));
         }
     }
-
-
-    public static void serverTick(Level level, BlockPos pos, BlockState state, FreezerBlockEntity blockEntity) {
-        AbstractFurnaceBlockEntityAccessor abstractFurnaceBlockEntityAccessor = (AbstractFurnaceBlockEntityAccessor) blockEntity;
-        boolean flag = abstractFurnaceBlockEntityAccessor.callIsLit();
-        boolean flag1 = false;
-
-        if (abstractFurnaceBlockEntityAccessor.callIsLit()) {
-            abstractFurnaceBlockEntityAccessor.rubinatedNether$setLitTime(abstractFurnaceBlockEntityAccessor.rubinatedNether$getLitTime() - 1);
-        }
-
-        ItemStack itemstack = abstractFurnaceBlockEntityAccessor.rubinatedNether$getItems().get(1);
-        boolean flag2 = !abstractFurnaceBlockEntityAccessor.rubinatedNether$getItems().get(0).isEmpty();
-        boolean flag3 = !itemstack.isEmpty();
-        if (abstractFurnaceBlockEntityAccessor.callIsLit() || flag3 && flag2) {
-            Recipe<?> recipe;
-            if (flag2) {
-                recipe = abstractFurnaceBlockEntityAccessor.rubinatedNether$getQuickCheck().getRecipeFor(blockEntity, level).orElse(null);
-            } else {
-                recipe = null;
-            }
-
-            int i = blockEntity.getMaxStackSize();
-            if (!abstractFurnaceBlockEntityAccessor.callIsLit() && abstractFurnaceBlockEntityAccessor.callCanBurn(level.registryAccess(), recipe, abstractFurnaceBlockEntityAccessor.rubinatedNether$getItems(), i)) {
-                abstractFurnaceBlockEntityAccessor.rubinatedNether$setLitTime(abstractFurnaceBlockEntityAccessor.callGetBurnDuration(itemstack));
-                abstractFurnaceBlockEntityAccessor.rubinatedNether$setLitDuration(abstractFurnaceBlockEntityAccessor.rubinatedNether$getLitTime());
-                if (abstractFurnaceBlockEntityAccessor.callIsLit()) {
-                    flag1 = true;
-                    if (itemstack.hasCraftingRemainingItem())
-                        abstractFurnaceBlockEntityAccessor.rubinatedNether$getItems().set(1, itemstack.getCraftingRemainingItem());
-                    else
-                    if (flag3) {
-                        itemstack.shrink(1);
-                        if (itemstack.isEmpty()) {
-                            abstractFurnaceBlockEntityAccessor.rubinatedNether$getItems().set(1, itemstack.getCraftingRemainingItem());
-                        }
-                    }
-                }
-            }
-
-            if (abstractFurnaceBlockEntityAccessor.callIsLit() && abstractFurnaceBlockEntityAccessor.callCanBurn(level.registryAccess(), recipe, abstractFurnaceBlockEntityAccessor.rubinatedNether$getItems(), i)) {
-                abstractFurnaceBlockEntityAccessor.rubinatedNether$setCookingProgress(abstractFurnaceBlockEntityAccessor.rubinatedNether$getCookingProgress() + 1);
-                if (abstractFurnaceBlockEntityAccessor.rubinatedNether$getCookingProgress() == abstractFurnaceBlockEntityAccessor.rubinatedNether$getCookingTotalTime()) {
-                    abstractFurnaceBlockEntityAccessor.rubinatedNether$setCookingProgress(0);
-                    abstractFurnaceBlockEntityAccessor.rubinatedNether$setCookingTotalTime(AbstractFurnaceBlockEntityAccessor.callGetTotalCookTime(level, blockEntity));
-                    if (blockEntity.burn(level.registryAccess(), recipe, blockEntity.items, i)) {
-                        blockEntity.setRecipeUsed(recipe);
-                    }
-
-                    flag1 = true;
-                }
-            } else {
-                abstractFurnaceBlockEntityAccessor.rubinatedNether$setCookingProgress(0);
-            }
-        } else if (!abstractFurnaceBlockEntityAccessor.callIsLit() && abstractFurnaceBlockEntityAccessor.rubinatedNether$getCookingProgress() > 0) {
-            abstractFurnaceBlockEntityAccessor.rubinatedNether$setCookingProgress(Mth.clamp(abstractFurnaceBlockEntityAccessor.rubinatedNether$getCookingProgress() - 2, 0, abstractFurnaceBlockEntityAccessor.rubinatedNether$getCookingTotalTime()));
-        }
-
-        if (flag != abstractFurnaceBlockEntityAccessor.callIsLit()) {
-            flag1 = true;
-            state = state.setValue(AbstractFurnaceBlock.LIT, abstractFurnaceBlockEntityAccessor.callIsLit());
-            level.setBlock(pos, state, 1 | 2);
-        }
-
-        if (flag1) {
-            setChanged(level, pos, state);
-        }
-
-        if (abstractFurnaceBlockEntityAccessor.rubinatedNether$getItems().get(0).isEmpty() && abstractFurnaceBlockEntityAccessor.rubinatedNether$getItems().get(2).isEmpty()) {
-            blockEntity.remainderItem = ItemStack.EMPTY; // Resets the remainder item variable used for hopper extraction at the end of the tick loop. This is necessary so that it actually has enough time to get extracted.
-        }
-    }
-
-    /**
-     * Ensures that NBT is carried over between input and result stacks.<br><br>
-     * Warning for "unchecked" is suppressed because casting {@link Recipe}<{@link WorldlyContainer}> is fine and done by vanilla.
-     * @param recipe The {@link Recipe Recipe<?>} being burned.
-     * @param stacks The {@link NonNullList NonNullList<ItemStack>} of items in the menu.
-     * @param stackSize The max stack size as an {@link Integer}.
-     * @return A {@link Boolean} for whether the item successfully burnt.
-     */
-    @SuppressWarnings("unchecked")
-    private boolean burn(RegistryAccess registryAccess, @Nullable Recipe<?> recipe, NonNullList<ItemStack> stacks, int stackSize) {
-        AbstractFurnaceBlockEntityAccessor abstractFurnaceBlockEntityAccessor = (AbstractFurnaceBlockEntityAccessor) this;
-        if (recipe != null && abstractFurnaceBlockEntityAccessor.callCanBurn(registryAccess, recipe, stacks, stackSize)) {
-            ItemStack inputSlotStack = stacks.get(0);
-            ItemStack resultStack = ((Recipe<WorldlyContainer>) recipe).assemble(this, registryAccess);
-            ItemStack resultSlotStack = stacks.get(2);
-
-            if (inputSlotStack.is(resultStack.getItem())) {
-                EnchantmentHelper.setEnchantments(EnchantmentHelper.getEnchantments(inputSlotStack), resultStack);
-                if (inputSlotStack.hasTag()) {
-                    resultStack.setTag(inputSlotStack.getTag());
-                }
-            }
-            if (inputSlotStack.is(resultStack.getItem())) {
-                resultStack.setDamageValue(0);
-            }
-
-            if (resultSlotStack.isEmpty()) {
-                stacks.set(2, resultStack.copy());
-            } else if (resultSlotStack.is(resultStack.getItem())) {
-                resultSlotStack.grow(resultStack.getCount());
-            }
-
-            if (inputSlotStack.hasCraftingRemainingItem() && !inputSlotStack.getCraftingRemainingItem().is(resultStack.getCraftingRemainingItem().getItem())) {
-                stacks.set(0, inputSlotStack.getCraftingRemainingItem());
-            } else {
-                inputSlotStack.shrink(1);
-            }
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * Allows the Aether's furnaces to have remaining crafting byproducts extracted (like buckets) alongside product items.
-     * @param index The {@link Integer} for the slot index.
-     * @param stack The {@link ItemStack} trying to be taken from the block.
-     * @param direction The {@link Direction} for a face.
-     * @return Whether the item can be taken by a hopper through the face, as a {@link Boolean}.
-     */
-    @Override
-    public boolean canTakeItemThroughFace(int index, ItemStack stack, Direction direction) {
-        AbstractFurnaceBlockEntityAccessor abstractFurnaceBlockEntityAccessor = (AbstractFurnaceBlockEntityAccessor) this;
-        Optional<NonNullList<Ingredient>> ingredient = abstractFurnaceBlockEntityAccessor.rubinatedNether$getQuickCheck().getRecipeFor(this, this.level).map(AbstractCookingRecipe::getIngredients);
-        if (this.remainderItem.isEmpty()) {
-            ingredient.ifPresent(ing -> this.remainderItem = stack.getCraftingRemainingItem()); // Stores the correlating crafting remainder item.
-        }
-        if (direction == Direction.DOWN && index == 0) {
-            if (!this.remainderItem.isEmpty()) {
-                return stack.is(this.remainderItem.getItem()); // An item can be taken as long as it matches the stored crafting remainder item.
-            } else {
-                return false;
-            }
-        } else {
-            return true;
-        }
-    }
-
-    @Override
-    public int[] getSlotsForFace(Direction direction) {
-        if (direction == Direction.DOWN) {
-            return SLOTS_FOR_DOWN;
-        } else {
-            return direction == Direction.UP ? SLOTS_FOR_UP : SLOTS_FOR_SIDES;
-        }
-    }
-
-    @Override
-    public boolean canPlaceItem(int index, ItemStack stack) {
-        if (index == 2) {
-            return false;
-        } else if (index != 1) {
-            return true;
-        } else {
-            return this.getBurnDuration(stack) > 0;
-        }
-    }
-
 }
