+++ src/main/java/net/artienia/rubinated_nether/screen/FreezerFuelSlot.java
@@ -1,19 +1,28 @@
 package net.artienia.rubinated_nether.screen;
 
 import net.minecraft.world.Container;
+import net.minecraft.world.inventory.AbstractFurnaceMenu;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 
 public class FreezerFuelSlot extends Slot {
-    private FreezerMenu menu;
+    private final FreezerMenu menu;
 
-    public FreezerFuelSlot(FreezerMenu menu, Container pContainer, int pSlot, int pX, int pY) {
-        super(pContainer, pSlot, pX, pY);
-        this.menu = menu;
+    public FreezerFuelSlot(FreezerMenu pFurnaceMenu, Container pFurnaceContainer, int pSlot, int pXPosition, int pYPosition) {
+        super(pFurnaceContainer, pSlot, pXPosition, pYPosition);
+        this.menu = pFurnaceMenu;
     }
 
-    @Override
-    public boolean mayPlace(ItemStack stack) {
-        return this.menu.isFuel(stack);
+    public boolean mayPlace(ItemStack pStack) {
+        return this.menu.isFuel(pStack) || isBucket(pStack);
+    }
+
+    public int getMaxStackSize(ItemStack pStack) {
+        return isBucket(pStack) ? 1 : super.getMaxStackSize(pStack);
+    }
+
+    public static boolean isBucket(ItemStack pStack) {
+        return pStack.is(Items.BUCKET);
     }
 }
