+++ src/main/resources/data/rubinated_nether/recipes/ruby_from_molten_ruby.json
@@ -1,7 +1,7 @@
 {
   "type": "rubinated_nether:freezing",
   "category": "freezable_misc",
-  "cookingtime": 500,
+  "cookingtime": 400,
   "experience": 1.0,
   "ingredient": {
     "item": "rubinated_nether:molten_ruby"
